Add support for persistent volumes

+++ cmd/kubernetes-performance/main.go
@@ -4,29 +4,36 @@
 package main
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"io"
 	"log"
 	"os"
 	"path/filepath"
+	"strings"
 	"time"
 
 	"github.com/svent/go-flags"
 	apiv1 "k8s.io/api/core/v1"
+	"k8s.io/apimachinery/pkg/api/resource"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/client-go/kubernetes"
+	_ "k8s.io/client-go/plugin/pkg/client/auth"
 	"k8s.io/client-go/tools/clientcmd"
 )
 
 var options struct {
-	Namespace  string `long:"namespace" default:"kubernetes-performance" description:"Namespace for the workload"`
-	KubeConfig string `long:"kube-config" env:"KUBECONFIG" default:"" description:"The location of the Kubernetes configuration"`
-	Command    string `long:"command" default:"" description:"The command to execute"`
+	Namespace    string `long:"namespace" default:"kubernetes-performance" description:"Namespace for the workload"`
+	KubeConfig   string `long:"kube-config" env:"KUBECONFIG" default:"" description:"The location of the Kubernetes configuration"`
+	Command      string `long:"command" default:"sysbench cpu run --time=30 --threads=4" description:"The command to execute"`
+	Cleanup      bool   `long:"cleanup" default:"false" description:"Cleanup pods after run"`
+	Pvc          bool   `long:"pvc" default:"" description:"Claim a persistent volume and mount to the pods"`
+	StorageClass string `long:"storage-class" default:"standard" description:"Persistent volume storage class"`
 }
 
 func main() {
+	var fsGroup = int64(1000)
+
 	args, err := flags.Parse(&options)
 	if err != nil {
 		if et, ok := err.(*flags.Error); ok {
@@ -76,21 +83,66 @@ func main() {
 	fmt.Printf("There are %d pods in the namespace\n", len(pods.Items))
 
 	for _, node := range nodes.Items {
+		if options.Pvc {
+			pvc := &apiv1.PersistentVolumeClaim{
+				ObjectMeta: metav1.ObjectMeta{
+					Name: fmt.Sprintf("kubernetes-performance-%s", node.Name),
+				},
+				Spec: apiv1.PersistentVolumeClaimSpec{
+					StorageClassName: &options.StorageClass,
+					AccessModes:      []apiv1.PersistentVolumeAccessMode{apiv1.ReadWriteOnce},
+					Resources: apiv1.ResourceRequirements{
+						Requests: apiv1.ResourceList{
+							"storage": resource.MustParse("1Gi"),
+						},
+					},
+				},
+			}
+
+			_, err := clientset.CoreV1().PersistentVolumeClaims(options.Namespace).Create(context.TODO(), pvc, metav1.CreateOptions{})
+			if err != nil {
+				panic(err.Error())
+			}
+		}
+
 		pod := &apiv1.Pod{
 			ObjectMeta: metav1.ObjectMeta{
 				Name: fmt.Sprintf("kubernetes-performance-%s", node.Name),
 			},
 			Spec: apiv1.PodSpec{
 				NodeName: node.Name,
+				SecurityContext: &apiv1.PodSecurityContext{
+					FSGroup: &fsGroup,
+				},
 				Containers: []apiv1.Container{
 					{
 						Name:    "kubernetes-performance",
-						Image:   "nginx:1.12",
-						Command: []string{"echo", "1"},
+						Image:   "registry.gitlab.com/delta10/kubernetes-performance:latest",
+						Command: strings.Split(options.Command, " "),
 					},
 				},
 				RestartPolicy: apiv1.RestartPolicyNever,
 			},
+		}
+
+		if options.Pvc {
+			pod.Spec.Containers[0].VolumeMounts = []apiv1.VolumeMount{
+				{
+					Name:      "pvc",
+					MountPath: "/pvc",
+				},
+			}
+
+			pod.Spec.Volumes = []apiv1.Volume{
+				{
+					Name: "pvc",
+					VolumeSource: apiv1.VolumeSource{
+						PersistentVolumeClaim: &apiv1.PersistentVolumeClaimVolumeSource{
+							ClaimName: fmt.Sprintf("kubernetes-performance-%s", node.Name),
+						},
+					},
+				},
+			}
 		}
 
 		_, err := clientset.CoreV1().Pods(options.Namespace).Create(context.TODO(), pod, metav1.CreateOptions{})
@@ -111,12 +163,12 @@ func main() {
 
 		podsCompleted = true
 		for _, pod := range pods.Items {
-			if pod.Status.Phase == apiv1.PodPending {
+			if pod.Status.Phase == apiv1.PodPending || pod.Status.Phase == apiv1.PodRunning {
 				podsCompleted = false
 			}
 		}
 
-		if podsCompleted != true {
+		if podsCompleted == false {
 			fmt.Printf("Waiting for pods to complete...\n")
 			time.Sleep(5 * time.Second)
 		} else {
@@ -137,17 +189,37 @@ func main() {
 		}
 		defer podLogs.Close()
 
-		buf := new(bytes.Buffer)
-		_, err = io.Copy(buf, podLogs)
+		logFile, err := os.Create(fmt.Sprintf("%s.log", pod.Name))
 		if err != nil {
 			panic(err.Error())
 		}
 
-		fmt.Printf("Logs %s: %s\n", pod.Name, buf.String())
+		defer logFile.Close()
 
-		err = clientset.CoreV1().Pods(options.Namespace).Delete(context.TODO(), pod.Name, metav1.DeleteOptions{})
+		_, err = io.Copy(logFile, podLogs)
 		if err != nil {
 			panic(err.Error())
+		}
+
+		if options.Cleanup {
+			err = clientset.CoreV1().Pods(options.Namespace).Delete(context.TODO(), pod.Name, metav1.DeleteOptions{})
+			if err != nil {
+				panic(err.Error())
+			}
+		}
+	}
+
+	if options.Cleanup {
+		pvcs, err := clientset.CoreV1().PersistentVolumeClaims(options.Namespace).List(context.TODO(), metav1.ListOptions{})
+		if err != nil {
+			panic(err.Error())
+		}
+
+		for _, pvc := range pvcs.Items {
+			err = clientset.CoreV1().PersistentVolumeClaims(options.Namespace).Delete(context.TODO(), pvc.Name, metav1.DeleteOptions{})
+			if err != nil {
+				panic(err.Error())
+			}
 		}
 	}
 }
